An interface with multiple IP addresses is internally represented by multiple interfaces with exactly one IP address. These additional interfaces are called secondary interfaces. In previous versions, if an interface was referenced in a rule, all IP addresses are used. Now, only the first (primary) IP address is used. You have to add the secondary interface to its IP address used.
A Secondary interface by default gets a name which is derived from the
name of the primary interface by adding an incrementing number
beginning with "2".  E.g. interface:router.name.2,
interface:router.name.3 .

This naming scheme is used, if a list of ip addresses is specified for
an interface. Alternatively, individual names can be given by a new syntax:

interface:name = { ip = 1.2.3.4;
                   secondary:name2 = { ip = 1.2.3.99; }
}