Fixed setup of {is_identical} relatonship. Previously this could get wrong with multiple nat networks and one loopback interface having all the same IP address.
Fixed {up} relation during local_optimization.
Nat networks identical to a loopback network must be larger than the loopback
network. Otherwise the corresponding loopback interface wouldn't be recognized
as identical to the nat network.

Only generate secondary rule during local_optimization, if such a rule does'nt
exist already. Normally this doesn't happen, because the original rule would
be optimized away. But if the original src or dst is expanded to a network
which in turn is modified to {is_identical}, we need this new test.