Better check and optimize networks in subnet relation.
The notion of 'any object' has been split to 'security zone' and
'aggregate'.

A security zone or in short, zone is contiguous part of the topology
with managed devices at its border and only unmanaged devices inside
the zone.

An object 'any:X' or 'any:[...]' aggregates networks inside a zone.
Until now, an aggregate matches all networks inside a zone.  This
version introduces aggregates with ip/mask, matching only networks
inside the zone, which are a subnet of ip/mask.

Attributes 'owner' and 'no_in_acl' for a zone can still be set with an
aggregate having no mask or mask 0.

This version has much more checks and optimizations for networks A < B
in subnet relation, where both networks are located inside the same
zone.

- It checks and warns for redundant rules for A and B.

- A inherits the owner of B.

- Checks for missing any rules have been extended to all supernets,
  i.e. all networks having one or more subnets.  For each rule having
  a supernet as source or destination, the path from source to
  destination is checked.  If some zone on the path contains networks
  Ni matching the supernet, it is checked, that the same rule is also
  defined for a supernet or aggregate of all Ni. If only a single N1
  matches, it is sufficient to have the same rule with N1.

Netspoc gives a warning for an undeclared subnet relation.
It can be declared with attribute 'subnet_of' at subnet.
Previously, a supernet could be declared with attribute 'route_hint'.
This has been replaced by attribute 'has_subnets'.
The restrictions of 'route_hint' are gone.
A supernet can have hosts now and can freely be used in rules.