Use largest supernet in zone for routing.
Add largest supernet inside the zone, if available.
This is needed, because we use the supernet in secondary optimization too.
Moreover this reduces the number of routing entries.
It isn't sufficient to solely use the supernet because network and supernet
can have different next hops at end of path.
For an aggregate, take all matching networks inside the zone.
These are supernets by design.