Fixed bug with multiple aggregates having same IP.
The bug occured if you defined different matching aggregates
any:X = { link = network:Y; ip = i.i.i.i/n; }
- with mask != 0
- having the same IP address
- located in different security zones.
In this case the subnet relation to the enclosing aggregate with mask = 0
was only calculated for one matching aggregate.
As a subsequent fault, for one or more enclosing aggregates, the list of
subnets was missing the subnets of the matching aggregate.