Automatically add managed hosts to rule.
We recently added support for managed hosts.  A rule which permits
access to network, implictly permits access to hosts of this
network. Until now, access to each managed host had to be permitted
explicitly. This was counter intuitive.

We now implicitly permit access to contained managed hosts, if a rule
permits traffic to a network or aggregate.