Fixed route for subnet < net having different hops
Be networks N1 < N2 < N3 in subnet relation.
Static route for N1, N3 has next hop H1,
N2 has hop H2.
Route to N1 must not be removed as redundant to N3.
Otherwise, traffic to N1 would be routed to H2.

This is also observed if N3 is a default route.