Extended valid names and handling of ID hosts.
Previously, the name needed to contain a '@' and
was always checked against the 'ea' field of the certificate.
Now a name without '@' is also valid. It is checked against some
other subject-name of the certificate.
Use the new attribute 'check-subject-name' inside 'radius_attributes'
to configure the to be checked subject-name.