Move attribute src_range from protocol to rule.
This reverts 7d29183 and fixes #11.

src_range is only stored, if we have a real source port or source port
range.
If no src_range is given, it defaults to protocol "ip".
We no longer use "tcp 1-65535" or "udp 1-65535" as source range,
but use implicit value "ip" in this case.
Splitted protocols are already expanded in expand_protocols.

Changes:
- Protocol having modifiers is no longer valid in "general_permit".

Fixes:
- Protocols of "general_permit" are respected for Linux devices now.