Optimized processing for rules at managed=local.
Up to this version, rules for devices with managed=local had been
distributed to routers and removed later during local_optimization.

Now, networks and aggregates are analyzed before, if they are filtered
at managed=local or not. So we can efficiently decide during rules
distribution, if a rule is needed or not.

This can give a huge performance improvement, if managed=local is in use.

As an additional benefit, supernet rules, that ar no longer
distributed, are also not checked any longer. This fixes #3.