Better secondary optimization
Check for disabling secondary optimization is more precise now.
Optimization is only disabled if destination is same network.
With rules
1. permit any->net:dst1, telnet
2. permit host:src->host:dst2, http
rule 2 is still optimized, even if host:dst2 is in same zone than net:dst1.

Optimization for rule with multiple interfaces of a single router is
more restricted now. It is only applied if
- the destination router itself is "managed = secondary" or
- destination has only one real and optionally multiple loopback interfaces.